requirements.txt
alembic==1.14.1
babel==2.18.0
blinker==1.8.2
click==8.1.8
dnspython==2.6.1
dominate==2.9.1
email-validator==2.3.0
Flask==3.0.3
flask-babel==3.1.0
Flask-Bootstrap==3.3.7.1
Flask-DotEnv==0.1.2
Flask-Login==0.6.3
Flask-Mail==0.10.0
Flask-Migrate==4.1.0
Flask-Moment==1.0.6
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
greenlet==3.1.1
guess-language-spirit==0.5.3
idna==3.11
importlib_metadata==8.5.0
importlib_resources==6.4.5
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==2.1.5
packaging==26.0
PyJWT==2.9.0
PyMySQL==1.1.2
python-dotenv==1.0.1
pytz==2025.2
SQLAlchemy==2.0.45
typing_extensions==4.13.2
visitor==0.1.3
Werkzeug==3.0.6
WTForms==3.1.2
zipp==3.20.2
